moves all helper methods out of sales_analyst
